+++ src/mouse_tracking/cli/infer.py
