+++ src/mouse_tracking/cli/main.py
