+++ src/mouse_tracking/cli/utils.py
